fix: Code style improvements
