+++ update-claims-features/github.go
@@ -338,3 +338,26 @@ func (m *UpdateClaimsFeatures) prExists(ctx context.Context, branchName string) 
 
 	return nil, nil
 }
+
+func (m *UpdateClaimsFeatures) getReleases(ctx context.Context) (string, error) {
+	ghReleaseListResult, err := dag.Gh(dagger.GhOpts{
+		Version: m.GhCliVersion,
+	}).Container(dagger.GhContainerOpts{
+		Token: m.PrefappGhToken,
+		Repo:  "prefapp/features",
+	}).WithDirectory(m.ClaimsDirPath, m.ClaimsDir, dagger.ContainerWithDirectoryOpts{}).
+		WithWorkdir(m.ClaimsDirPath).
+		WithEnvVariable("CACHE_BUSTER", time.Now().String()).
+		WithExec([]string{
+			"gh",
+			"release",
+			"list",
+			"--limit",
+			"999",
+			"--json",
+			"name",
+		}).
+		Stdout(ctx)
+
+	return ghReleaseListResult, err
+}
